Payement_app/Backend/api/ai_module/enquiry_ai/enquiry_ai_requirements.txt
langchain_core==0.3.79
langchain_google_genai==2.1.12
langchain_postgres==0.0.15
langgraph==0.6.10
pydantic==2.12.3
python_pptx==1.0.2
SQLAlchemy==2.0.44
python-dotenv==1.1.1
grpcio==1.62.2
grpcio-status==1.62.2